feat: typescript rewrite
