+++ .gitignore
@@ -7,3 +7,5 @@ node_modules
 .DS_Store
 coverage
 dist
+sw.*
+.env
