--- commitlint.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  extends: [
-    '@commitlint/config-conventional'
-  ]
-}
