--- husky.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  hooks: {
-    'commit-msg': 'commitlint -E HUSKY_GIT_PARAMS',
-    'pre-commit': 'yarn lint',
-    'pre-push': 'yarn lint'
-  }
-}
