+++ jest.config.js
@@ -1,5 +1,4 @@
 module.exports = {
-  collectCoverage: true,
-  collectCoverageFrom: ['lib/**/*.js'],
-  testEnvironment: 'node'
+  preset: '@nuxt/test-utils',
+  collectCoverageFrom: ['src/**']
 }
