--- lib/logger.js
@@ -1,4 +1,0 @@
-
-const consola = require('consola')
-
-module.exports = consola.withScope('nuxt:netlify-files')
