+++ package.json
@@ -5,33 +5,38 @@
   "repository": "nuxt-community/netlify-files-module",
   "license": "MIT",
   "contributors": [
-    {
-      "name": "Alexander Lichter <npm@lichter.io>"
-    }
+    "Alexander Lichter <npm@lichter.io>",
+    "Ricardo Gobbo de Souza <ricardogobbosouza@yahoo.com.br>"
   ],
-  "main": "lib/module.js",
+  "main": "./dist/module.js",
+  "types": "./dist/module.d.ts",
   "files": [
-    "lib"
+    "dist"
   ],
   "scripts": {
-    "lint": "eslint --ext .js,.vue .",
-    "release": "yarn test && standard-version && git push --follow-tags && npm publish",
-    "test": "yarn lint && jest"
+    "build": "siroc build",
+    "prepublishOnly": "yarn build",
+    "lint": "eslint --ext .js,.ts,.vue .",
+    "release": "yarn test && yarn build && standard-version && git push --follow-tags && npm publish",
+    "test": "yarn lint && yarn jest"
   },
   "dependencies": {
     "@iarna/toml": "^2.2.5",
     "consola": "^2.15.0",
+    "defu": "^3.2.2",
     "node-html-parser": "^3.0.3"
   },
   "devDependencies": {
-    "@commitlint/cli": "latest",
-    "@commitlint/config-conventional": "latest",
-    "@nuxtjs/eslint-config": "latest",
-    "@nuxtjs/module-test-utils": "latest",
+    "@babel/preset-typescript": "latest",
+    "@nuxt/test-utils": "latest",
+    "@nuxt/types": "latest",
+    "@nuxtjs/eslint-config-typescript": "latest",
+    "@types/jest": "latest",
+    "@types/node": "latest",
     "eslint": "latest",
-    "husky": "latest",
     "jest": "latest",
-    "nuxt-edge": "latest",
+    "nuxt": "latest",
+    "siroc": "latest",
     "standard-version": "latest"
   },
   "publishConfig": {
