+++ test/fixture/warn/nuxt.config.js
@@ -1,6 +1,6 @@
-module.exports = {
+export default {
   rootDir: __dirname,
   buildModules: [
-    { handler: require('../../../') }
+    '../../../src/module.ts'
   ]
 }
