+++ test/utils.ts
@@ -1,6 +1,7 @@
-const { join, resolve } = require('path')
+import { join, resolve } from 'path'
+import { NuxtOptions } from '@nuxt/types'
 
-const resolvePaths = (nuxt) => {
+export const resolvePaths = (nuxt: { options: NuxtOptions }) => {
   const distBasePath = resolve(nuxt.options.rootDir, nuxt.options.generate.dir)
   const headersPath = join(distBasePath, '_headers')
   const redirectsPath = join(distBasePath, '_redirects')
@@ -13,8 +14,4 @@ const resolvePaths = (nuxt) => {
     tomlPath,
     formsPath
   }
-}
-
-module.exports = {
-  resolvePaths
 }
